numpy 1.12.0rc2 compatibility fix related to variable typecasting

+++ theano/sparse/tests/test_sp2.py
@@ -62,7 +62,7 @@ class PoissonTester(utt.InferShapeTester):
 
 
 class BinomialTester(utt.InferShapeTester):
-    n = tensor.scalar()
+    n = tensor.scalar(dtype='int64')
     p = tensor.scalar()
     shape = tensor.lvector()
     _n = 5
